Move UDB2EventServicesProvider out of Silex namespace
